+++ HISTORY.md
@@ -39,6 +39,9 @@
     * German
     * Chinese, thanks, Simon!
 
+### 2021-10-12
+* scenes, store: store single palette setting per project (for making extensions)
+
 ### 2021-10-11
 * objects: sort order of blocks in custom categories alphabetically in the unified palette